requirements.txt
PyQt6
jinja2
pyinstaller
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
